split the volumes related code to volumes.go
